+++ source/index.ts
@@ -8,18 +8,41 @@ export interface LineInfo {
 	file: string
 }
 
+/**
+ * If provided, continue skipping until:
+ *
+ * 1. The file or method is found
+ * 2. Once found, will continue until neither the file nor method are found anymore
+ * 3. Once exited, the frame offset will then apply
+ *
+ * If you wish to capture the method or the file, combine them with `frames: -1`.
+ *
+ * If you wish for more customisation than this, create an issue requesting passing a custom skip handler function, as more variance to this interface is too much customisation complexity.
+ */
 export interface LineOffset {
-	/** continue skipping frames until we encounter this method */
+	/**
+	 * if provided, continue until a method containing this string is exited
+	 * if provided alongside a file, will continue until neither the file nor method are found
+	 * this allows file and method to act as fallbacks for each other, such that if one is not found, it doesn't skip everything
+	 */
 	method?: string | null
-	/** continue skipping frames until we encounter this file */
+	/**
+	 * if provided, continue until a file containing this string is exited
+	 * if provided alongside a method, will continue until neither the file nor method are found
+	 * this allows file and method to act as fallbacks for each other, such that if one is not found, it doesn't skip everything
+	 */
 	file?: string | null
-	/** once we have encountered our first desired frame, continue for this many frames */
+	/**
+	 * once we have satisfied the found condition, if any, then apply this index offset to the frames
+	 * e.g. 1 would mean next frame, and -1 would mean the previous frame
+	 * Use -1 to go back to the found method or file
+	 */
 	frames?: number
 }
 
 /**
  * Get the information about the line that called this method.
- * @param offset set this to the distance between this method and the true caller
+ * @param offset continue until these offset conditions are met, used to continue to the caller
  * @throws if a failure occured creating the line info
  * @example Input
  * ``` javascript
@@ -36,7 +59,11 @@ export interface LineOffset {
  * ```
  */
 export default function getCurrentLine(
-	offset: LineOffset = { file: __filename, method: 'getCurrentLine', frames: 1 }
+	offset: LineOffset = {
+		file: __filename,
+		method: 'getCurrentLine',
+		frames: 0,
+	}
 ): LineInfo {
 	// prepare
 	const result: LineInfo = {
@@ -83,39 +110,50 @@ export default function getCurrentLine(
 			.filter((line: string) => line.length !== 0)
 
 		// Continue
-		let foundFile: boolean = !offset.file
-		let foundMethod: boolean = !offset.method
+		let found: boolean = !offset.file && !offset.method
+		let exited: boolean = found
 
 		// Parse our lines
-		for (const line of lines) {
-			// offset
-			if (!foundFile && line.includes(offset.file)) {
-				foundFile = true
-			}
-			if (!foundMethod && line.includes(offset.method)) {
-				foundMethod = true
-			}
-			if (!foundFile || !foundMethod) {
-				continue
-			} else if (offset.frames) {
-				--offset.frames
+		let i = 0
+		while (i < lines.length) {
+			const line = lines[i]
+
+			// continue
+			if (found && exited) {
+				const parts = line.split(':')
+				if (parts.length >= 2) {
+					if (parts[0].indexOf('(') === -1) {
+						result.method = 'unknown'
+						result.file = parts[0].replace(/^.+?\s+at\s+/, '')
+					} else {
+						result.method = parts[0]
+							.replace(/^.+?\s+at\s+/, '')
+							.replace(/\s+\(.+$/, '')
+						result.file = parts[0].replace(/^.+?\(/, '')
+					}
+					result.line = Number(parts[1])
+					break
+				}
+				// found and exited, but not a valid entry, so continue
+				++i
 				continue
 			}
 
-			// extract
-			const parts = line.split(':')
-			if (parts.length >= 2) {
-				if (parts[0].indexOf('(') === -1) {
-					result.method = 'unknown'
-					result.file = parts[0].replace(/^.+?\s+at\s+/, '')
-				} else {
-					result.method = parts[0]
-						.replace(/^.+?\s+at\s+/, '')
-						.replace(/\s+\(.+$/, '')
-					result.file = parts[0].replace(/^.+?\(/, '')
-				}
-				result.line = Number(parts[1])
-				break
+			// continue until found and exited
+			if (line.includes(offset.file) || line.includes(offset.method)) {
+				found = true
+				// next item
+				++i
+				continue
+			} else if (found) {
+				// exited, apply frame offset and call it a day
+				i += offset.frames || 0
+				exited = true
+				continue
+			} else {
+				// nothing found yet, next item
+				++i
+				continue
 			}
 		}
 	} catch (err) {